Use DateTimeStyles.AdjustToUniversal.
Using DateTimeStyles.AdjustToUniversal instead of DateTimeStyles.None returns a correct but converted-to-Local time (not Unspecified) if a Utc time was saved (which would have caused a Z to be appended). (Behavior if a Local or Unspecified time was saved is not changed by DateTimeStyles.AdjustToUniversal, i.e. it still returns Unspecified in both cases.)